+++ temporalio/contrib/strands/__init__.py
@@ -2,12 +2,14 @@
 
 from . import workflow
 from ._plugin import StrandsPlugin
+from ._sandbox_activity import SandboxWorkflowContext
 from ._temporal_agent import TemporalAgent
 from ._temporal_mcp_client import TemporalMCPClient
 from ._temporal_sandbox import TemporalSandbox
 
 __all__ = [
     "StrandsPlugin",
+    "SandboxWorkflowContext",
     "TemporalAgent",
     "TemporalMCPClient",
     "TemporalSandbox",